+++ geoopt/__init__.py
@@ -28,4 +28,4 @@ from .manifolds import (
     BirkhoffPolytope,
 )
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"