+++ geoopt/optim/radam.py
@@ -87,11 +87,11 @@ class RiemannianAdam(OptimMixin, torch.optim.Adam):
                     exp_avg = state["exp_avg"]
                     exp_avg_sq = state["exp_avg_sq"]
                     # actual step
-                    grad.add_(weight_decay, point)
+                    grad.add_(point, alpha=weight_decay)
                     grad = manifold.egrad2rgrad(point, grad)
-                    exp_avg.mul_(betas[0]).add_(1 - betas[0], grad)
+                    exp_avg.mul_(betas[0]).add_(grad, alpha=1 - betas[0])
                     exp_avg_sq.mul_(betas[1]).add_(
-                        1 - betas[1], manifold.component_inner(point, grad)
+                        manifold.component_inner(point, grad), alpha=1 - betas[1]
                     )
                     if amsgrad:
                         max_exp_avg_sq = state["max_exp_avg_sq"]
@@ -120,7 +120,10 @@ class RiemannianAdam(OptimMixin, torch.optim.Adam):
                     exp_avg.set_(exp_avg_new)
 
                     group["step"] += 1
-                if self._stabilize is not None and group["step"] % self._stabilize == 0:
+                if (
+                    group["stabilize"] is not None
+                    and group["step"] % group["stabilize"] == 0
+                ):
                     self.stabilize_group(group)
         return loss