+++ geoopt/optim/rsgd.py
@@ -93,13 +93,13 @@ class RiemannianSGD(OptimMixin, torch.optim.Optimizer):
                     else:
                         manifold = self._default_manifold
 
-                    grad.add_(weight_decay, point)
+                    grad.add_(point, alpha=weight_decay)
                     grad = manifold.egrad2rgrad(point, grad)
                     if momentum > 0:
                         momentum_buffer = state["momentum_buffer"]
-                        momentum_buffer.mul_(momentum).add_(1 - dampening, grad)
+                        momentum_buffer.mul_(momentum).add_(grad, alpha=1 - dampening)
                         if nesterov:
-                            grad = grad.add_(momentum, momentum_buffer)
+                            grad = grad.add_(momentum_buffer, alpha=momentum)
                         else:
                             grad = momentum_buffer
                         # we have all the things projected
@@ -114,7 +114,10 @@ class RiemannianSGD(OptimMixin, torch.optim.Optimizer):
                         copy_or_set_(point, new_point)
 
                     group["step"] += 1
-                if self._stabilize is not None and group["step"] % self._stabilize == 0:
+                if (
+                    group["stabilize"] is not None
+                    and group["step"] % group["stabilize"] == 0
+                ):
                     self.stabilize_group(group)
         return loss